ginthi_agents/etl_service/requirements.txt
fastapi
uvicorn
playwright
pydantic
python-dotenv
aiofiles
